requirements.txt
attrs==24.2.0
audioread==3.0.1
azure-core==1.31.0
azure-storage-blob==12.23.1
boto3==1.35.41
botocore==1.35.41
boxsdk==3.13.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.1
decorator==5.1.1
greenlet==3.1.1
idna==3.10
isodate==0.7.2
jmespath==1.0.1
joblib==1.4.2
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
msgpack==1.1.0
numba==0.60.0
numpy==2.0.2
packaging==24.1
pandas==2.2.3
platformdirs==4.3.6
pooch==1.8.2
pycparser==2.22
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-toolbelt==1.0.0
s3transfer==0.10.3
scikit-learn==1.5.2
scipy==1.13.1
six==1.16.0
soundfile==0.12.1
soxr==0.5.0.post1
SQLAlchemy==2.0.36
threadpoolctl==3.5.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
